Refactored Address Service Test
Moved it to the actual service as not the ressource is tested.
